modeling

+++ rapidstream_script/pipeline.py
@@ -1,6 +1,6 @@
 from rapidstream import PipelineConfig
 
 config = PipelineConfig(
-    pp_scheme="single",
+    pp_scheme="double",
 )
 config.save_to_file("pipeline_config.json")
